--- src/main/java/racingcar/utils/CarsGenerator.java
@@ -1,14 +1,0 @@
-package racingcar.utils;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import racingcar.domain.Car;
-
-public class CarsGenerator {
-
-    public static List<Car> generateCars(List<String> carNames) {
-        return carNames.stream()
-                .map(Car::new)
-                .collect(Collectors.toList());
-    }
-}
